requirements.txt
python-dotenv
pywhatkit
openai
